extract static course into a component
