+++ pix-editor/app/components/form/static-course.hbs
@@ -1,0 +1,58 @@
+<form {{on "submit" this.onSubmitClicked}} class="form">
+  <Card class="new-static-course-form-field" @title="1. Renseigner les informations">
+    <PixInput
+      @id="static-course-name"
+      @label="Nom du test statique"
+      @errorMessage="{{this.name.errorMessage}}"
+      @requiredLabel="{{this.name.errorMessage}}"
+      @validationStatus={{this.name.state}}
+      @value={{@initialName}}
+      {{on "keyup" this.updateName}}
+      {{on "focusout" this.validateName}}
+    />
+    <PixTextarea
+      @id="static-course-description"
+      @label="Description à usage interne"
+      @maxlength="1000"
+      @value={{@initialDescription}}
+      {{on "keyup" this.updateDescription}}
+    />
+  </Card>
+  <Card class="new-static-course-form-card" @title="2. Sélectionner les épreuves">
+    <span>
+      Veuillez renseigner les IDs que vous souhaitez ajouter à votre test statique, une par ligne dans l'ordre souhaité. Le premier ID renseigné correspondra à l'ID de la première épreuve du test statique.
+    </span>
+    <PixTextarea
+      @id="static-course-challenges"
+      @label="IDs des épreuves"
+      @maxlength="1000"
+      rows="20"
+      @value={{@initialChallengeIds}}
+      @errorMessage="{{if this.challengeIds.isError this.challengeIds.errorMessage ''}}"
+      {{on "keyup" this.updateChallengeIds}}
+      {{on "focusout" this.validateChallengeIds}}
+    />
+  </Card>
+  <div class="form-error">
+    {{#each this.errorMessages as |errorMessage|}}
+      {{errorMessage}}<br/>
+    {{/each}}
+  </div>
+  <div class="page-actions">
+    <PixButton
+      @backgroundColor="transparent-light"
+      @isBorderVisible={{true}}
+      @triggerAction={{@onFormCancelled}}
+    >
+      {{@cancelButtonText}}
+    </PixButton>
+    <PixButton
+      @backgroundColor="blue"
+      @type="submit"
+      @isDisabled={{this.isFormInvalid}}
+      @isLoading={{this.isSubmitting}}
+    >
+      {{@submitButtonText}}
+    </PixButton>
+  </div>
+</form>
