+++ pix-editor/app/templates/authenticated/static-courses/list.hbs
@@ -10,7 +10,6 @@
           <PixButtonLink
             @backgroundColor="blue"
             @route="authenticated.static-courses.new"
-            @iconBefore="plus"
             @isDisabled={{not this.model.mayCreateStaticCourse}}
             aria-describedby='create-static-course-tooltip'
             class="pix-button-link-with-icon white-font"
