+++ pix-editor/app/templates/authenticated/static-courses/new.hbs
@@ -3,60 +3,14 @@
 </header>
 <main class="page-body">
   <section class="page-section">
-    <form {{on "submit" this.createStaticCourse}} class="form">
-      <Card class="new-static-course-form-field" @title="1. Informations">
-        <PixInput
-          @id="static-course-name"
-          @label="Nom"
-          @errorMessage="{{this.name.errorMessage}}"
-          @requiredLabel="{{this.name.errorMessage}}"
-          @validationStatus={{this.name.state}}
-          {{on "keyup" this.updateName}}
-          {{on "focusout" this.validateName}}
-        />
-        <PixTextarea
-          @id="static-course-description"
-          @label="Description à usage interne"
-          @maxlength="1000"
-          {{on "keyup" this.updateDescription}}
-        />
-      </Card>
-      <Card class="new-static-course-form-card" @title="2. Sélection des épreuves">
-        <span>
-          Renseignez les IDs que vous souhaitez ajouter à votre test statique. Un ID par ligne dans l'ordre souhaité.
-        </span>
-        <PixTextarea
-          @id="static-course-challenges"
-          @label="Épreuves"
-          @maxlength="1000"
-          rows="20"
-          @errorMessage="{{if this.challengeIds.isError this.challengeIds.errorMessage ''}}"
-          {{on "keyup" this.updateChallengeIds}}
-          {{on "focusout" this.validateChallengeIds}}
-        />
-      </Card>
-      <div class="form-error">
-        {{#each this.errorMessages as |errorMessage|}}
-          {{errorMessage}}<br/>
-        {{/each}}
-      </div>
-      <div class="page-actions">
-        <PixButtonLink
-          @backgroundColor="transparent-light"
-          @route="authenticated.static-courses.list"
-          @isBorderVisible={{true}}
-        >
-          Annuler
-        </PixButtonLink>
-        <PixButton
-          @backgroundColor="blue"
-          @type="submit"
-          @isDisabled={{this.isFormInvalid}}
-          @isLoading={{this.isSubmitting}}
-        >
-          Créer le test statique
-        </PixButton>
-      </div>
-    </form>
+    <Form::StaticCourse
+      @initialName=""
+      @initialDescription=""
+      @initialChallengeIds=""
+      @cancelButtonText="Annuler"
+      @onFormCancelled={{this.goBackToList}}
+      @submitButtonText="Créer le test statique"
+      @onFormSubmitted={{this.createStaticCourse}}
+    />
   </section>
 </main>
